Changed from create-react-app to a Nextjs application (Experimental if this is last commit).
