+++ next-app/src/app/Styles/login.css
@@ -1,0 +1,128 @@
+@import url('https://fonts.googleapis.com/css2?family=Abril+Fatface&display=swap');
+
+:root {
+    --mainColor: #6999EB;
+    --mainColorLight: #798eea;
+    --mainColorDark: #05386B;
+    --mainColorDarker: #032445;
+    --secondaryColor: #4ACA85;
+    --secondaryColorDark: #3bb473;
+    --tertiaryColor: #FFD700;
+    --textColor: #EDF5E1;
+}
+
+
+.login-container {
+    display: flex;
+    justify-content: center;
+}
+
+
+.login-panel {
+    width: 400px;
+    margin-top: 200px;
+    border: 5px solid #ccc;
+    border-radius: 5px;
+    box-shadow: 2px 2px 15px #ccc;
+    padding: 30px;
+    backdrop-filter: blur(30px);
+    background-color: var(--mainColorDarker);
+    
+}
+  
+.title {
+    text-align: center;
+    margin-bottom: 30px;
+    color: var(--textColor);
+    font-family: 'Abril Fatface', cursive;
+    font-size: 45px;
+}
+
+
+.input-group {
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    margin: 30px 0 5px 0;
+}
+
+  
+.input-group label {
+    font-weight: bold;
+    margin-bottom: 10px;
+    color: var(--textColor);
+    font-size: larger;
+}
+  
+.input-group input {
+    padding: 10px;
+    font-size: 14px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 18px;
+}
+
+i {
+    position: absolute;
+    top: 25%;
+    right: 5%;
+
+}
+
+i:hover{
+    cursor: pointer;
+    color: var(--secondaryColorDark);
+}
+  
+button[type="submit"] {
+    margin: 30px 0 50px;
+    width: 100%;
+    padding: 10px;
+    background-color: #55c2da;
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    font-size: 16px;
+    cursor: pointer;
+    font-weight: bold;
+    font-size: larger;
+}
+
+.border {
+    background-color: #ccc;
+    height: 3px;
+    box-shadow: 1px 1px 6px #ccc;
+}
+
+button[type="submit"]:hover{
+   background-color: #3c9aaf;
+   box-shadow: 1px 1px 30px #3c9aaf;
+}
+
+p {
+    margin: 5px;
+    color: var(--textColor);
+    float: right;
+}
+
+button#new-account[type="submit"] {
+    background-color: var(--secondaryColor);
+    margin-top: 40px;
+    margin-bottom: 5px;
+
+}
+
+button#new-account[type="submit"]:hover {
+    background-color: var(--secondaryColorDark);
+    box-shadow: 1px 1px 30px var(--secondaryColorDark);
+}
+
+
+a.forgot-password {
+    float: right;
+    color: var(--textColor);
+}
+
+a.forgot-password:hover{
+    color: var(--secondaryColor);
+}
