+++ next-app/src/app/globals.css
@@ -1,0 +1,269 @@
+@import url('https://fonts.googleapis.com/css2?family=Lato&display=swap');
+
+@import url('https://fonts.googleapis.com/css2?family=Abril+Fatface&display=swap');
+
+
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+    font-family: 'Lato', sans-serif;
+}
+
+
+:root {
+    --mainColor: #6999EB;
+    --mainColorLight: #798eea;
+    --mainColorDark: #05386B;
+    --mainColorDarker: #032445;
+    --secondaryColor: #4ACA85;
+    --secondaryColorDark: #379683;
+    --tertiaryColor: #FFD700;
+    --textColor: #EDF5E1;
+
+    --graphBG: url('..\\Images\\stock-background-4608x3456.jpg')
+}
+
+
+#sign-up-btn {
+    margin-left: 25px;
+    margin-right: 25px;
+    width: 100px;
+
+    color: var(--textColor);
+    font-size: 18px;
+
+    box-shadow: 1px 1px 20px var(--mainColor);
+}
+
+#sign-up-btn:hover {
+    background-color: var(--secondaryColor);
+    box-shadow: 1px 1px 30px var(--secondaryColor);
+    border: 1px solid var(--secondaryColor);
+}
+
+.bg-mainColor {
+    background-color: rgba(5, 56, 107, 0.99);
+}
+
+.bg-secondaryColor {
+    background-color: var(--secondaryColor);
+}
+
+
+
+.Navbar .logo {
+    color: var(--textColor);
+    display: flex;
+    justify-content: left;
+    align-items: center;
+}
+
+.Navbar .logo img {
+    width: 96px;
+    float: left;
+}
+
+.Navbar .logo h3 {
+    float: right;
+    padding-left: 1rem;
+    color: var(--secondaryColor);
+    font-family: 'Abril Fatface', cursive;
+    font-size: 35px;
+}
+
+.Navbar .login-button {
+    width: 64px;
+    height: 64px;
+    margin-left: 100px;
+}
+
+.Navbar .login-button img:hover{
+    filter: sepia(100%) hue-rotate(120deg) saturate(900%) brightness(250%) drop-shadow(0px 0px 20px var(--secondaryColor));
+    cursor: pointer;
+}
+
+.Navbar .login-button img {
+    width: 100%;
+    filter: invert(100%);
+    position: relative;
+    height: 64px;
+    
+}
+
+.Navbar .navbar {
+    padding: 0 2rem 0 2rem;
+    margin-bottom: 77px;
+}
+
+.Navbar .navbar-nav {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.Navbar .navbar-nav a {
+    color: var(--textColor);
+    margin: 0 2rem;
+    font-size: larger;
+    font-weight: bold;
+}
+
+.Navbar .navbar-nav a:hover {
+    color: var(--secondaryColor);
+    text-shadow: 0px 0px 30px var(--secondaryColor);
+}
+
+.Navbar .navbar-nav a.active {
+    color: var(--secondaryColor);
+}
+
+
+header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    height: 80px;
+    padding: 0 2rem;
+    background-color: var(--mainColor);
+    color: var(--textColor);
+}
+
+header .nav-btn {
+    padding: 5px;
+    cursor: pointer;
+    background: transparent;
+    border: none;
+    outline: none;
+    color: var(--textColor);
+    visibility:  hidden;
+    opacity: 0;
+    font-size: 1.8rem
+}
+
+#mobile-collapse-btn {
+    position: absolute;
+    right: 20px;
+    top: 25px;
+}
+
+@media only screen and (max-width: 768px) {
+    .Navbar .login-button {
+        visibility: hidden;
+    }
+}
+
+
+.login-container {
+    display: flex;
+    justify-content: center;
+}
+
+
+.login-panel {
+    width: 400px;
+    margin-top: 200px;
+    border: 5px solid #ccc;
+    border-radius: 5px;
+    box-shadow: 2px 2px 15px #ccc;
+    padding: 30px;
+    backdrop-filter: blur(30px);
+    
+}
+  
+.title {
+    text-align: center;
+    margin-bottom: 30px;
+    color: var(--textColor);
+    font-family: 'Abril Fatface', cursive;
+    font-size: 45px;
+}
+
+
+.input-group {
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    margin: 30px 0 5px 0;
+}
+
+  
+.input-group label {
+    font-weight: bold;
+    margin-bottom: 10px;
+    color: var(--textColor);
+    font-size: larger;
+}
+  
+.input-group input {
+    padding: 10px;
+    font-size: 14px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 18px;
+}
+
+i {
+    position: absolute;
+    top: 25%;
+    right: 5%;
+
+}
+
+i:hover{
+    cursor: pointer;
+    color: var(--secondaryColorDark);
+}
+  
+button[type="submit"] {
+    margin: 30px 0 50px;
+    width: 100%;
+    padding: 10px;
+    background-color: #55c2da;
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    font-size: 16px;
+    cursor: pointer;
+    font-weight: bold;
+    font-size: larger;
+}
+
+.border {
+    background-color: #ccc;
+    height: 3px;
+    box-shadow: 1px 1px 6px #ccc;
+}
+
+button[type="submit"]:hover{
+   background-color: #3c9aaf;
+   box-shadow: 1px 1px 30px #3c9aaf;
+}
+
+p {
+    margin: 5px;
+    color: var(--textColor);
+    float: right;
+}
+
+button#new-account[type="submit"] {
+    background-color: var(--secondaryColor);
+    margin-top: 40px;
+    margin-bottom: 5px;
+
+}
+
+button#new-account[type="submit"]:hover {
+    background-color: var(--secondaryColorDark);
+    box-shadow: 1px 1px 30px var(--secondaryColorDark);
+}
+
+
+a.forgot-password {
+    float: right;
+    color: var(--textColor);
+}
+
+a.forgot-password:hover{
+    color: var(--secondaryColor);
+}
