+++ next-app/src/pages/index.js
@@ -1,0 +1,12 @@
+import Image from 'next/image';
+import { Inter } from '@next/font/google';
+
+import MyNavbar from '../app/Components/navbar';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export default function Home() {
+  return (
+    <MyNavbar/>
+  )
+}
