+++ next-app/src/pages/login.js
@@ -1,0 +1,8 @@
+import Login from "../app/Components/login";
+
+export default function LoginPage() {
+    return (
+      <Login/>
+    )
+  }
+
