--- react-app/src/Pages/404.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-import "./PageStyles/404page.css";
-
-function NoPage() {
-    return (
-        <div className="nopage">
-            <h2>404 bruh this url dont exist</h2>
-        </div>
-    )
-}
-
-
-export default NoPage;
