--- react-app/src/Pages/Home.jsx
@@ -1,30 +1,0 @@
-import React from "react";
-
-import MyNavbar from "../Components/navbar";
-
-import "./PageStyles/homepage.css";
-
-function Home() {
-    return (
-        <div className="Homepage">
-            <MyNavbar/>
-            <div className="container">
-              <p>Start basic text here</p>
-            </div>
-        </div>
-    )
-}
-
-// When the user scrolls down 20px from the top of the document, slide down the navbar
-// When the user scrolls to the top of the page, slide up the navbar (50px out of the top view)
-window.onscroll = function() {scrollFunction()};
-
-function scrollFunction() {
-  if (document.body.scrollTop > 20 || document.documentElement.scrollTop > 20) {
-    document.getElementById("navbar").style.top = "0";
-  } else {
-    document.getElementById("navbar").style.top = "-50px";
-  }
-}
-
-export default Home;
