--- react-app/src/Pages/PageStyles/404page.css
@@ -1,5 +1,0 @@
-.nopage {
-    display: flex;
-    justify-content: center;
-    margin: 100px;
-}
