--- react-app/src/Pages/PageStyles/homepage.css
@@ -1,20 +1,0 @@
-
-
-.Homepage {
-    background-image: url("../../Images/stock-background-4608x3456.jpg");
-    width: 100vw;
-    height: 100vh;
-    scroll-behavior: smooth;
-}
-
-.container {
-    justify-content: center;
-    align-items: center;
-    display: flex;
-}
-
-p{
-    text-align: left;
-    padding-top: 80px;
-    color: whitesmoke;
-}
