--- react-app/src/Pages/PageStyles/loginpage.css
@@ -1,8 +1,0 @@
-.Loginpage {
-    background-color: #032445;
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    overflow: hidden;
-}
