--- react-app/src/Pages/PageStyles/signuppage.css
@@ -1,9 +1,0 @@
-.Signup-page {
-    background-color: #032445;
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    overflow: hidden;
-    scroll-behavior: smooth;
-}
