--- react-app/src/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import './Styles/index.css';
-
-
-
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
-);
